vaak-lite/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
groq>=0.4.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
